+++ publisher-sdk/src/main/java/com/criteo/publisher/tasks/WebViewDataTask.java
@@ -103,7 +103,7 @@ public class WebViewDataTask extends SafeRunnable {
       runOnUiThreadExecutor.executeAsync(new SafeRunnable() {
         @Override
         public void runSafely() {
-          listener.onAdReadyToDisplay();
+          listener.onAdReceived();
         }
       });
     }