+++ CHANGELOG.md
@@ -14,6 +14,9 @@
   - Move `CriteoInterstitialAdDisplayListener#onAdFailedToDisplay(CriteoErrorCode)` to
    `CriteoInterstitialAdListener#onAdFailedToReceive(CriteoErrorCode)`. This callback is fired
     when an error happens after requesting an interstitial ad.
+  - Move `CriteoInterstitialAdDisplayListener#onAdReadyToDisplay()` to
+   `CriteoInterstitialAdListener#onAdReceived()`. This callback is fired when an interstitial ad
+    is ready to be displayed.
 
 # Version 3.9.0